Move BibtexString to model package
