+++ src/main/java/net/sf/jabref/collab/ChangeScanner.java
@@ -38,6 +38,7 @@ import net.sf.jabref.imports.ParserResult;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.BibtexDatabase;
 import net.sf.jabref.model.BibtexEntry;
+import net.sf.jabref.model.BibtexString;
 import net.sf.jabref.util.Util;
 
 public class ChangeScanner implements Runnable {
