+++ src/main/java/net/sf/jabref/collab/StringChange.java
@@ -21,6 +21,7 @@ import javax.swing.JScrollPane;
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.BibtexDatabase;
+import net.sf.jabref.model.BibtexString;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
