+++ src/main/java/net/sf/jabref/export/FileActions.java
@@ -31,10 +31,7 @@ import net.sf.jabref.*;
 import net.sf.jabref.gui.GUIGlobals;
 import net.sf.jabref.logic.config.SaveOrderConfig;
 import net.sf.jabref.logic.l10n.Localization;
-import net.sf.jabref.model.BibtexDatabase;
-import net.sf.jabref.model.BibtexEntry;
-import net.sf.jabref.model.BibtexEntryType;
-import net.sf.jabref.model.CustomEntryType;
+import net.sf.jabref.model.*;
 
 public class FileActions {
