+++ src/main/java/net/sf/jabref/gui/undo/UndoableInsertString.java
@@ -21,6 +21,7 @@ import net.sf.jabref.*;
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.BibtexDatabase;
+import net.sf.jabref.model.BibtexString;
 
 public class UndoableInsertString extends AbstractUndoableEdit {
