+++ src/main/java/net/sf/jabref/imports/AppendDatabaseAction.java
@@ -37,6 +37,7 @@ import net.sf.jabref.gui.undo.UndoableInsertString;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.BibtexDatabase;
 import net.sf.jabref.model.BibtexEntry;
+import net.sf.jabref.model.BibtexString;
 import net.sf.jabref.util.Util;
 
 /**
