+++ src/main/java/net/sf/jabref/imports/ImportMenuItem.java
@@ -38,6 +38,7 @@ import net.sf.jabref.gui.undo.UndoableRemoveEntry;
 import net.sf.jabref.model.BibtexDatabase;
 import net.sf.jabref.model.BibtexEntry;
 import net.sf.jabref.model.BibtexEntryType;
+import net.sf.jabref.model.BibtexString;
 import net.sf.jabref.util.Util;
 
 /*
