+++ src/main/java/net/sf/jabref/sql/exporter/DBExporter.java
@@ -35,7 +35,7 @@ import javax.swing.JOptionPane;
 import net.sf.jabref.model.BibtexDatabase;
 import net.sf.jabref.model.BibtexEntry;
 import net.sf.jabref.model.BibtexEntryType;
-import net.sf.jabref.BibtexString;
+import net.sf.jabref.model.BibtexString;
 import net.sf.jabref.gui.JabRefFrame;
 import net.sf.jabref.MetaData;
 import net.sf.jabref.groups.structure.*;
