+++ src/main/java/net/sf/jabref/sql/importer/DBImporter.java
@@ -27,6 +27,7 @@ import net.sf.jabref.groups.GroupTreeNode;
 import net.sf.jabref.model.BibtexDatabase;
 import net.sf.jabref.model.BibtexEntry;
 import net.sf.jabref.model.BibtexEntryType;
+import net.sf.jabref.model.BibtexString;
 import net.sf.jabref.sql.DBImporterExporter;
 import net.sf.jabref.sql.DBStrings;
 import net.sf.jabref.sql.SQLUtil;
