+++ src/main/java/net/sf/jabref/wizard/auximport/AuxSubGenerator.java
@@ -64,6 +64,7 @@ import java.util.regex.Pattern;
 import net.sf.jabref.*;
 import net.sf.jabref.model.BibtexDatabase;
 import net.sf.jabref.model.BibtexEntry;
+import net.sf.jabref.model.BibtexString;
 
 public class AuxSubGenerator
 {
